edited duos tests without so many sleeps
