+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/test/duosTest/DuosLinkSpec.scala
@@ -1,5 +1,7 @@
 package org.broadinstitute.dsde.firecloud.test.duosTest
 
+import java.util.concurrent.TimeUnit
+
 import org.broadinstitute.dsde.firecloud.fixture.UserFixtures
 import org.broadinstitute.dsde.firecloud.page.duos.{DuosHomePage, DuosLoginPage}
 import org.broadinstitute.dsde.firecloud.page.{AuthenticatedPage, BaseFireCloudPage, PageUtil}
@@ -26,46 +28,35 @@ class DuosLinkSpec extends FreeSpec with ParallelTestExecution with Matchers
       withSignInDuos(user) { homePage =>
         homePage.datasetSearch()
         val datarequest = cssSelector("div[class='main-title-description']")
-        //        find(datarequest).get.underlying.getText should include ("Data Access Committee")
         find(datarequest).get.text should include("Data Access Committee")
         println("success")
       }
     }
   }
 
+    "clicking help and about button" in {
+          withWebDriver { implicit driver =>
+            goTo("https://duos.dsde-dev.broadinstitute.org/#/home")
 
-  //  "verify that the description of the DUOS graphic contains the attribute name" in {
-  //        withWebDriver { implicit driver =>
-  //          val duosp = CssSelectorQuery("img[alt*='What is DUOS graphic']")
-  //          val duosGraphic = new DuosLoginPage("https://duos.dsde-dev.broadinstitute.org/#/home").open.find()
-  //
-  //          val duosGraphic = new DuosLoginPage("https://duos.dsde-dev.broadinstitute.org/#/home").open
-  //            //val duosGraphic = new DuosLoginPage().open.duosP()
-  //
-  //            duosGraphic.underlying.isDisplayed shouldBe(true)
-  //          }
-  //        }
-  //
-  //  "click on the join DUOS button and type username in" in {
-  //    withWebDriver { implicit driver =>
-  //
-  //      val join = new DuosSignInPage().open.joinDuos()
-  //      // AuthenticatedPage.readUserEmail
-  //      join shouldBe("test.firec@gmail.com")
-  //    }
-  //  }
-  //
-  //  "clicking if the help button exists" in {
-  //        withWebDriver { implicit driver =>
-  //          val duosHelp = new DuosSignInPage().helpDuos()
-  //            duosHelp.underlying.isEnabled shouldBe(true)
-  //          }
-  //        }
-  //
-  //  "testing if the about button exists" in {
-  //        withWebDriver { implicit driver =>
-  //          val duosAbout = new DuosSignInPage().open.aboutDuos()
-  //          duosAbout.underlying.isEnabled shouldBe(true)
-  //          }
-  //        }
+            val duosp = find(CssSelectorQuery("img[alt*='What is DUOS graphic']"))
+            duosp.get.isDisplayed shouldBe(true)
+
+            val helpButton = find(CssSelectorQuery("a[href='#/home_help']"))
+            helpButton.get.isEnabled shouldBe (true)
+
+            val aboutButton = find(CssSelectorQuery("a[href='#/home_about']"))
+            aboutButton.get.isEnabled shouldBe(true)
+
+            val email = "test.firec@gmail.com"
+            val joinButton = CssSelectorQuery("a.navbar-duos-link-join")
+
+            find(joinButton).get.underlying.click()
+            Thread.sleep(1000)
+            val typeInDes = CssSelectorQuery("input[ng-model='form.name']")
+            find(typeInDes).get.asInstanceOf[ValueElement].value_=(email)
+
+            val fullnametext = cssSelector("label[class='home-control-label col-lg-12 col-md-12 col-sm-12 col-xs-12']")
+            find(fullnametext).get.text shouldBe("Full Name")
+            }
+          }
 }
